+++ src/app/models/StatusResponse.tsx
@@ -1,0 +1,5 @@
+export interface StatusResponse {
+  host: string;
+  running: boolean;
+  disks: string[];
+}
